requirements.txt
black==21.11b1
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
flake8==4.0.1
idna==3.3
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
pika==1.2.0
platformdirs==2.4.0
pycentral==0.0.3
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.19.2
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
tomli==1.2.2
typing-extensions==4.0.1
urllib3==1.26.7
